services/backend/requirements.txt
# Django
django==3.0.3
djangorestframework==3.11.0
